refactoring login form
