+++ app/components/views/GetStartedPage/RPCLoginForm/index.js
@@ -1,0 +1,8 @@
+import React, { Component } from "react";
+import {LoginRPCHeader, LoginFormBody} from "./LoginForm"
+
+class LoginForm extends Component {
+  
+}
+
+export { LoginRPCHeader, LoginFormBody };
